fix(ethercat): move slave-name dedup helper to frontend/utils [DOPE-281]
The helper was misplaced under backend/shared/ethercat/, which made the
store import violate the layer rule "Store must not import from Backend
Shared". Move it to frontend/utils/ alongside next-name.ts and
ethercat-status.ts — the conventional location for pure helpers shared
between store and components.
